SMB: Extract SMB, SMBExploiter, SMBPropagator into separate modules
