+++ monkey/agent_plugins/exploiters/smb/src/smb_propagation_client.py
@@ -1,0 +1,29 @@
+import logging
+
+from common.credentials import Credentials
+from common.event_queue import IAgentEventPublisher
+from infection_monkey.i_puppet import TargetHost
+
+logger = logging.getLogger(__name__)
+
+
+class SMBPropagationClient:
+    """Manages the RPC/SMB connection, Propagation events"""
+
+    def __init__(self, agent_event_publisher: IAgentEventPublisher):
+        self._agent_event_publisher = agent_event_publisher
+
+    def propagate(self, host: TargetHost, path: str, credentials: Credentials) -> bool:
+        """Do all of the propagation stuff"""
+
+        if not self._run_agent(host, path):
+            logger.debug(f"Failed to run agent on {host}")
+            return False
+
+        return True
+
+    def _run_agent(self, host, path: str):
+        # - Create RPC connection
+        # - Build agent run command
+        # - Use RPC to run the agent on the victim
+        pass
